:art: refactor sandbox for mutiple instance
